+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.3 Release (02.01.2021)
 * Added working leveling system [MySQL needs manual removal!]
+* Changed player death to 5 blocks out of arena instead of void
 
 ### 1.0.2 Release (01.01.2021)
 * Fixed Setup on 1.11 & 1.12